requirements.txt
PyGithub
openai
pytest
